requirements.txt
Flask==2.3.3
gunicorn==21.2.0
requests==2.31.0
newspaper3k==0.2.8
flask-limiter==3.5.0
python-decouple==3.8
lxml_html_clean==0.1.1
